added short formal variable names, some character limits
added short formal variable names, some character limits, and removed french ahead of english default/multi-language functionality.

short names are generated using the first two letters of the name of their table (or first letter of the first two words in a table name - i.e. the short formal name for a table) connected using an underscore to the first two letters or first letter of multiple words in the variable name. the variable name chunk varies from 2 to 4 characters in length.