Add uid column for unit and aggregation
Added a column for UID in unit.csv and aggregation.csv. These UIDs are three characters long, and for some are the same as the unitID/aggregationID. Meta-data on naming procedure to follow, but generally:

- UnidIDs that are 3 characters long or shorter have the same value as their UID
- single word unitIDs have the first three letters as their UID
-  For all others, the first letter of the first part of the UID, and then the first letter or digit of each chunk after were used up to a max of three characters.

For aggregation, the same all applies, but for multi-part aggregation IDs, the first two letters of the initial portion of the ID were used, and then the first letter of the next piece after that.