Simplified the get_dictionary function
The get_dictionary function used to assume that there would be multiple
dictionaries downloaded from OSF which is unnecessary. The
dictionary_path is now the full path to the downloaded dictionary rather
than a folder that could contain multiple dictionaries.

The download_dictionary function now returns the full path to the
dictionary rather than the folder to where it's downloaded.