Correcting Zones table issue
Corrected the mistake in the zones GitHub CSVs. The isoCode field should be 2 characters wide and contain only the country code, no zones codes. We have instances where zones are nested (to some extent) and so a zones code is then used in the isoCode column. The "-" and additional letters were removed from the isoCode column, and now only countries are present in that column.