Remove short names column, add as parts, create set
Part short names acting as a PK to to be references in the wideNames table were causing multiple indices in the SQL generation. Given that the short names were a minority, they were added instead as parts, and that column was cut. A new set was created to track these inputs and it was linked with the reportTableInput field.

Additionally, there was an incongruity in the "partLabel" column vs the partID entry and the ERD value - "label". So the column header was changed to "label".